Add VNET_BUFFER_F_AVAIL definitions
Add a check to make sure that the vlib and vnet buffer flag bit
definitions do not overlap.

The VNET_BUFFER_F_AVAIL1...8 definitions allow out-of-tree codes to:

  #define VNET_BUFFER_F_MY_USECASE VNET_BUFFER_F_AVAIL1

and so on. This avoids introducing irrelevant and/or proprietary bit
definitions into vnet/buffer.h, and hopefully minimizes merge pain for
everyone involved.

Change-Id: I5be4f61dceb81b5bfca005f6d609ade074af205b
Signed-off-by: Dave Barach <dave@barachs.net>